+++ Userland/Libraries/LibGfx/CMakeLists.txt
@@ -1,4 +1,5 @@
 include(skia)
+include(vulkan)
 
 set(SOURCES
     AffineTransform.cpp
@@ -65,6 +66,7 @@ set(SOURCES
     TextLayout.cpp
     Triangle.cpp
     VectorGraphic.cpp
+    SkiaBackendContext.cpp
 )
 
 serenity_lib(LibGfx gfx)
